Former-commit-id: d4e3581d70a74b1eda5236650d16b82774d55ebc

+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Events/Interview/AnswerDeclaredInvalid.cs
@@ -5,7 +5,7 @@ namespace WB.Core.SharedKernels.DataCollection.Events.Interview
 {
     public class AnswerDeclaredInvalid : QuestionPassiveEvent
     {
-        public AnswerDeclaredInvalid(Guid questionId, int[] propagationVector)
+        public AnswerDeclaredInvalid(Guid questionId, decimal[] propagationVector)
             : base(questionId, propagationVector) {}
     }
 }
